backend/requirement.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv

passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
